Rework the project admin pages
This commit completely reworks the project admin pages.
Prior to this commit, there was a single project admin page binding the
full project model to a single form. This made it difficult for users to
update project information and hard to maintain this part of the
application.

This commit uses ModelMapper to map only the relevant parts to the admin
UI and split the form into multiple pages.